Rename and refactor variable names.
